Image saving
images are converted to bitmap strings and saved in the respective note
