+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/SerializedImage.kt
@@ -3,8 +3,8 @@ package com.microsoft.device.display.samples.twonote
 import java.io.Serializable
 
 data class SerializedImage(
-    val authority: String,
-    val path: String,
+    val name: String,
+    val image: String,
     val coords: List<Float> = listOf(),
     val dimens: List<Int> = listOf()
 ) : Serializable
