+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/includes/Defines.kt
@@ -8,22 +8,23 @@
 object Defines {
     // File Handler references //
     const val DEFAULT_FILE_NAME = "drawing"
-    const val TEXT_PREFIX = "text/"
     const val IMAGE_PREFIX = "image/"
     const val PLAIN_TEXT = "text/plain"
+    const val TEXT_PREFIX = "text/"
 
-    // Drag constants //
-    const val RESIZE_SPEED = 15
+    // Image Handler constants //
     const val MIN_DIMEN = 250
+    const val RENDER_TIMER = 50L
+    const val RESIZE_SPEED = 15
     const val THRESHOLD = 0.2
 
     // Inking scaling constants //
-    const val PORT_TO_LAND = 4f / 3
     const val LAND_TO_PORT = 3f / 4
+    const val PORT_TO_LAND = 4f / 3
 
     // Fragment name constants //
-    const val LIST_FRAGMENT = "list fragment"
     const val DETAIL_FRAGMENT = "detail fragment"
-    const val NOTE = "note"
+    const val LIST_FRAGMENT = "list fragment"
     const val INODE = "inode"
+    const val NOTE = "note"
 }
